Consolidate ColorSelect2Content story
