+++ src/components/inputs/ColorSelect2/ColorSelect2.props.story.tsx
@@ -7,6 +7,10 @@ import { Button } from '../../Button/Button';
 
 import { Layout } from '../../../storybook';
 
+import { ColorSelect2PopoverContent } from './ColorSelect2PopoverContent';
+import { colorSpaces } from '../../../color';
+import { parseToHsl } from 'polished';
+
 export default {
   title: 'components/ColorSelect2/props',
 };
@@ -70,3 +74,57 @@ export function Disabled({ args }) {
     </Layout.StoryVertical>
   );
 }
+
+const Template = (args) => {
+  return (
+    <div
+      style={{
+        display: 'grid',
+        gap: '2rem',
+        gridTemplateColumns: 'repeat(2, min-content)',
+      }}
+    >
+      <ColorSelect2PopoverContent {...args} />
+      <ColorSelect2PopoverContent
+        {...args}
+        presets={{
+          palette: ['#909CDC', '#7BD8DB', '#78DD89'],
+          label: 'Presets',
+          onEdit: () => null,
+        }}
+      />
+      <ColorSelect2PopoverContent {...args} showHueSlider />
+      <ColorSelect2PopoverContent
+        {...args}
+        showHueSlider
+        presets={{
+          palette: ['#909CDC', '#7BD8DB', '#78DD89'],
+          label: 'Presets',
+          onEdit: () => null,
+        }}
+      />
+    </div>
+  );
+};
+
+const initialColors = colorSpaces(parseToHsl('#00adef'));
+const popoverContentProps = {
+  width: 260,
+  height: 200,
+  showHueSlider: false,
+  popOverRef: null,
+  presets: null,
+  dispatch: () => null,
+  onChange: () => null,
+  throttleSpeed: 40,
+  colorMeta: initialColors,
+  state: {
+    open: true,
+    editing: false,
+    error: false,
+    colorMeta: initialColors,
+    colorSpace: 'HEX',
+  },
+};
+export const PopoverContent = Template.bind({});
+PopoverContent.args = popoverContentProps;
